chore: upped sentry on `./api`
