+++ api/package.json
@@ -10,8 +10,8 @@
     "@dzcode.io/data": "*",
     "@dzcode.io/models": "*",
     "@dzcode.io/utils": "*",
-    "@sentry/node": "^7.46.0",
-    "@sentry/tracing": "^7.46.0",
+    "@sentry/node": "^8.28.0",
+    "@sentry/profiling-node": "^8.28.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.2",
     "class-validator-jsonschema": "^3.1.0",
